пошел по пути разных классов для хранимых объектов в двух уровнях понял что надо все сделать через интерфейы и перегрузку некоторых методов

+++ src/main/java/org/twolevelcache/MyObject.java
@@ -7,9 +7,9 @@ public class MyObject {
 
 
     @Getter
-    private Integer ID;
+    private Long ID;
 
-    public MyObject(Integer ObjectID){
+    public MyObject(Long ObjectID){
         ID = ObjectID;
     }
